Create models for post and comment.
