+++ comment/model/model.go
@@ -1,0 +1,7 @@
+package model
+
+type Comment struct {
+	ID     int64  `json:"id"`
+	Body   string `json:"body"`
+	PostID int64  `json:"postId"`
+}
