+++ post/model/model.go
@@ -1,0 +1,6 @@
+package model
+
+type Post struct {
+	ID    int64  `json:"id"`
+	Title string `json:"title"`
+}
